This is Sean's fence generating program that generates a basic chain-linked fence.  All parameters of the fence may be specified, and defaults may be changed through the header file in ../h.  The fence is an example of a procedural geometry generation.  The resulting fence easily contains a very large number of vectors.  The default fence generated consists of approx. 3 million vectors, but is only a 3meter x 2meter (w x h) fence.
Needless to say, the fence is detailed.  This type of generated geometry is
a good example of why being able to edit and view at different resolutions
would be a very good thing.  Currently, editing a geometry such as this is
not as simple a procedure as it could be.  Very long delays are encountered
while loading.

svn:revision:16033
cvs:account:morrison
cvs:branch:trunk