Added checking for problem matricies inside combinations.
It isn't good to have the rotation part of the matrix not be three
unit-length vectors forming an orthonormal basis. Modelers need to put
the (reciprocal of the) scale factor in the matrix[15] element, rather
than multiplying it along the primary diagonal in the rotation area.

svn:revision:17758
cvs:account:mike
cvs:branch:trunk