initial file.  adds initial support for the clone and tracker commands with updates to the code applied for changes that have occurred over the years.
these are the first of many contributions provided by the former GSI
with permission for inclusion in the open source release at last.  the
sources were found from an almost dead O2 external hard drive
enclosure.

thanks to Adam Ross for making the mods, to GSI for funding it and
ultimately returning the mods to the BRL-CAD community, and to Dwayne
Kregel for keeping up with the former GSI owners for permission and
with me to get the mods updated and added.

svn:revision:25772
cvs:account:brlcad
cvs:branch:trunk