The original problem was a result of the getint and getdouble functions using the actual field input size when parsing an input line and the bu_strlcpy macros in those two functions promptly chopped off the last character as they are supposed to do.  I replaced the macros with the bu_vls_strncpy functions so the parsing is more intuitive in that actual field size is entered and the last character is not chopped
The problem prevented successful conversion on any input deck, and
with the fix it now works on a complex TGM input file.

A successful round trip was also made with the same TGM.  Note that
"success" is within the constraints of the vdeck <-=> comgeom-g
converters where the solid, region, and group names are used
differently so one has to track the objects carefully to ensure the
trip was made correctly.

{above is from Tom Browder e-mail to brlcad-devel on 7 April 2012}

svn:revision:49960
svn:branch:trunk
svn:account:brlcad