Fix vlist drawing by nirt, rtcheck, etc. broken by r75470.
Fixing transparency exposed a problem in calls to invent_solid, all of which had "0.0" as the transparency argument. It's pretty clear the intention was to make temp vlists opaque (0% transparent), but confusingly we use "transparency" as a synonym for "alpha", so a transparency value of 0 actually means fully transparent.

svn:revision:75549
svn:branch:trunk
svn:account:n_reed