Making scripts better reflect conversion process
If at all avoidable, we don't want to have to manually update the custom commit
information again now that repowork allows for post-conversion email
manipulation.  However, we also want account-map to be accurate.  The secondary
repositories in the SVN repository (rt3, geomcore, etc.) will eventually be
converted with svn-fast-export - they do not use the custom process developed
for the main history.  Consequently, they will need a correct account-map file
with the most current addresses.

To handle this, we specifically stash the versions of the maps that match the
custom commits currently checked in and update the conversion scripts to reference
them specifically.  Also stash the email mapping file used by repowork, so we
can make sure we're capturing all updates made between the start of the conversion
and the final post-processing.

svn:revision:76465
svn:branch:trunk
svn:account:starseeker