Extract CVS commit info from cvs-fast-export
This script uses a cvs-fast-export conversion of the original CVS
repository without an authormap in order to extract the original
association between commit ids and commits.  Because SHA1s won't
map for the main conversion, use a combination of message SHA1
and timestamp to uniquely identify the commits.  When we move to
map this information into the final commit messages, we will
need to generate a similar map between this information and the
SHA1 values in the repowork input repository.  Reading both maps
will then serve as a key connecting new SHA1 values to the old
CVS data.

In addition, we also used the --first-parent option of rev-list
to follow the actual branch chains for the CVS conversions -
these will be recorded as cvs:branch: entries.  Anything not
on a branch will be deemed to be cvs:branch:trunk commits.

svn:revision:76554
svn:branch:trunk
svn:account:starseeker