skip the bigdb test if we can't allocate enough memory.
cmake 3.1 has a feature where a test will be considered skipped if
there's a particular exit code.  use this to detect when the bigdb
test is aborting due to not being able to allocate 5x the amount of
requested memory (this is a librt issue).

svn:revision:77713
svn:branch:trunk
svn:account:brlcad