Adjust GUI apps default working dir to HOME.
When launched from a system installed location via icon, the default
working directory of MGED, Archer and rtwizard gets set to the bin
directory that holds the executable.  This is typically a read-only
location, which results in a variety of problems when trying to run
various commands.

The most typical case where this is a problem is Windows desktop and
start menu shortcuts.  The NSIS installer has a way to adjust the
directory, but the MSI installer generated by WiX is a tougher nut to
crack.  Various experiments have so far failed to identify a working
approach to adjusting the shortcut startup directories with CMake/WiX.
Using Windows bat files as launchers works, but using them with the
MSI installer results in incorrect icons for the desktop shortcuts
with no obvious way to override them.

Rather than continue to jump through hoops, just set the working dir
directly to the user's home directory when these three applications
are invoked without arguments.  This should preserve all existing
behaviors for any scripting cases, but will avoid a read-only working
directory when using the no-args launch produced by desktop shortcuts.

svn:revision:78045
svn:branch:trunk
svn:account:starseeker