Rename - test updates using script succeeded.
To test the efficacy of this script, one of the updating
files to the conversion was altered to remove one of the
earliest correction lines, and that line was applied in
isolation to create a new repository with most of the
SHA1 ids altered.  The original repository and the
altered repository were then used to update the full input
files (minus the already applied line) by mapping the
original SHA1 keys to the corresponding new keys.  After
application of the altered files, the final git log --all
log from the resulting repository was compared to the
equivalent log generated from the repository created by
applying the original files to the original input. The logs
were identical, which means all the mapping were successfully
made and the equivalent changes successfully applied.

svn:revision:78313
svn:branch:trunk
svn:account:starseeker