Update fastgen regression LF/CRLF test handling
fast4 files classify as text files to VCS, so their line endings get
updated on checkout.  Normally this won't matter, but this particular
regression test is intended to make sure that it in fact does *not*
matter.  For that, we need to preserve the LF (UNIX) or CLRF (DOS) line
endings in the files upon checkout.

Rather than use .gitattributes (which invite mishaps if pattern matches
produce unintentional matches elsewhere in the tree or get stale as
files change) we instead compress the LF and CRLF versions of the files
to be checked out as tbz2 archives.  They are then decompressed when
needed.  We can do this since CMake itself is able to do the
decompression, and we are guaranteed to have CMake on any platform where
building is possible in the first place.

As an extra check, the regression script also verifies that the LF and
CRLF versions of the input are in fact different after being unpacked.