Stash the git conversion tkhtml repair scripts
The original git conversion of BRL-CAD deliberately removed most of the
expanded RCS tags from its source files, in order to make git's diff and
history tracking as efficient as possible.  However, it was missed that
the tkhtml build was actually aware of these populated tags and would
not compile without them.  This resulted in a situation where historical
checkouts would not compile due to tkhtml's not having its RCS tags,
even though the original forms of the trees would have compiled.

There were a couple of possible workarounds for this.  Git itself can
populate the $Id$ RCS tag, which happened to be the tag at issue -
although its output is nothing like other tools, it suffices to achieve
compilable sources.  This can be set up with a single line run from the
top level brlcad checkout:

echo "/src/other/tkhtml/** ident" > .git/info/attributes

A second option was to manually provide a workable change to the sources
after checkout:

find . -path \*tkhtml\* -type f -exec sed -i 's/\$Id\$/\$Id: tkhtml\$/g' {} \;

In the end, however, we found a way to correct the repository history
directly.  The scripts being committed here, along with a new option
added to the repowork tool (now being stored it its own github repo)
were used to generate fast-import files representing the original state
of the files in SVN, and alter the data in the repository by mapping the
old Git tkhtml file references to the SHA1 signatures of the SVN
versions.  As long as the blob fast-import file is loaded first into the
target repository so the blobs are there to reference, the fast-import
file written by repowork can successfully import and reference the
the original SVN files.