Document exec files in SVN history
It was found when attempting to perform distcheck on a Git checkout of
the rel-7-32-2 tag that not all files had preserved their executable
properties during the translation.  This manifested in a distcheck
failure where the run.sh script for benchmark, when copied to
bin/benchmark, failed to run.

To correct this issue, a repowork feature was added (see
https://github.com/BRL-CAD/repowork/commit/5f0cd70e0a2e2de153337bed87ae)
to update the modes of specified paths globally throughout the history,
and a list of SVN paths that at some point in the repository history had
been executables was generated.  Since this list (svn_exec_all.txt)
included a fair number of files with the flag accidentally set, Sean
audited the list and produced a smaller target set (svn_exec_final.txt).
Those paths were then set executable throughout the Git history, to
ensure they would run as expected if checked out.

This is something of a brute force solution in that it does not capture
at what points in the SVN history these flags were set or unset on
various files, but the cost of trying to more precisely identify and set
the executable mode in that fashion was deemed too high - setting the
flags globally for these paths addresses the functional issue, and was
relatively straightforward to achieve.