First cut at updating HACKING, COPYING, README
Undoubtedly a lot more to do here, but make an attempt to update the
HACKING workflow and other references to the SVN repository.

Note that the binary and source file hosting were not updated yet, and
the mailing lists are also unchanged as yet.  Those will be updated as
new solutions come online.