Identify the basic issues with ExtProject_Target
Three remaining issues surfaced with the RPATH setting - two minor, one
trickier.

Minor #1 - Itcl and Itk didn't have the RPATH flag set in their target
definitions (fixed).

Minor #2 - 3rd party binaries RPATHs don't match those produced by the
main CMake build - using bin instead of lib (investigating).

More significant is the behavior of CMake's RPATH_CHANGE - it is
correctly setting the final path, but it is not clearing the build dir
path.  So far I've not found a way to make the available commands in
CMake do the job, so I'm falling back on a solution similar to that
necessary on Apple and using the chrpath utility.  The code in this
commit works if a system chrpath is present - we'll need to bundle a
version in misc/tools to make sure we have the capabilities we need
reliably.  (chrpath is GPL, so we can't install it or use it as anything
except a build tool - unfortunately I've not found a BSD/MIT licensed
tool for this type of rpath work...)