Explain the reason for chrpath usage.
It's probably worth more study, but the odd library runpath in the
installed libgdal.so is originating from the GDAL build itself.  So far
it's looking like that path set is a consequence of what CMake is doing
internally to track RPATHs, and I've not yet identified a combination of
CMAKE_*RPATH* settings that avoids it.

We're getting the following from the GDAL_BLD ExternalProject_Add
installed version of the lib:

Library runpath: [/home/user/brlcad/build/lib::::::::::::::::::::::::::::::::::::/home/user/brlcad/build/lib]

when we should be getting (and are getting, prior to the GDAL_BLD install step):

Library runpath: [/home/user/brlcad/build/lib::::::::::::::::::::::::::::::::::::]

Then, when the parent CMake build does its install, the first part of
the path is overwritten but the appended path from GDAL_BLD persists.
The resulting binaries will run, but is an untidy state in which to
leave the runpaths.

I've not found a way yet using vanilla CMake to scrub down the final
results to what we want - simply ensuring chrpath is available to do the
job instead seems to be simpler all around.  (Only tested on Ubuntu
Linux so far - needs testing on the BSDs as well.)