Switch to using tclsh for pkgIndex generation
btclsh will attempt to use the generated pkgIndex files, and in a
parallel compilation it is possible for one of the files to be in an
intermediate state when btclsh tries to run to launch another one.  Use
tclsh here instead, since it won't try to load any of the files being
generated.