Need matrix from GED to be set in OpenGL
Going to need a fair bit more of the state from MGED to match its
visuals.  Setting the matrix gets us a basic view, but it's zoomed too
far out and the clipping plan is in the middle of the geometry.
Checkpoint anyway since this is our first wireframe visual in Qt with a
libdm OpenGL backend.