Experiment with creating a hash of bview
If it proves practical, a better solution than storing full copies of
the previous view state would be to create a hash.  The application can
then just store the previous view hash, and after command execution
check the new hash against the old to see if we need to draw.

This compiles, but isn't yet tested to see if it hashes correctly.