Opaque resizing seems to work better.
When moving the QSplitter up and down, was getting occasional crashes of
the form "Cannot make QOpenGLContext current in a different thread".
This is similar to problems encountered with Tk in rtwizard - opaque
resizing (where the window is continually redrawn as it is resized)
doesn't work so well for the OpenGL drawing context.

Fortunately, the same solution is available - just postpone the redraw
of the window contents until the user releases the splitter.