Add a context parameter to dm_open
Historically, dm_open took a Tcl interp pointer and deduced from the
Tcl/Tk context whether it needed a new window or not.  When operating
(potentially, not necessarily) without a Tcl/Tk interp, we need another
way to tell if we are being given an existing context or if we need to
create one.  With the ctx paramater, we can tell.