First translation of mouse view updating from MGED
Start unwinding MGED's interpretation of mouse motions - it does a lot
of different things depending on its mode.  We don't want modal
interfaces, but we're probably still going to need most of the mouse
behaviors to realize the same interface features.

The eventual idea is to translate the "post X,Y delta" calculations on
the bview into utility routines that are application and toolkit
agnostic.  This test implements view centered rotation of the view
within the camera.