Qt on OSX is going to need native OpenGL
OSX (unsurprisingly) is presenting issues.  First one is making sure we
find the right OpenGL to use with Qt - that means not mixing X11 and
Native OpenGL, which is going to be a problem for non-Aqua Tk, but
that's a problem for later...

More concerning is the qdm demo isn't displaying, even when we get the
libraries lined up.  The widget is present, since we're getting
printouts from mouse events, but the wireframe isn't being drawn...