Implement both single and rendering thread dmGL
For testing purposes, provide all-up implementations of dmGL both with
the rendering thread approach and with a paintGL override.  It's looking
like we can't mix the two in one widget.

Thus far, the rendering thread approach doesn't seem to work on the
Mac.  This particular iteration is unlikely to change that.  We'll see
about the single threaded version...