Rework OpenGL searching to be more general
This is still a work in progress (needs some more changes and lots of
cross platform testing) but the Qt work exposed a problem on OSX with
how we're handling OpenGL searching.

Historically, we only built with the X11 version of OpenGL on OSX, and
so had to constrain the find_package for OpenGL to avoid the native
OpenGL framework (which is normally what Mac packages would want.)  To
that end we have been using a FindGL.cmake file tuned to that specific
need.

However, when enabling Qt OpenGL applications on OSX, we're back to
wanting (needing, actually) the native OpenGL framework.  There will
probably be some work to get the Qt code working with the system
framework, but before we get to that point we need a way to pinpoint
which OpenGL library we need for which part of the code.  A global
solution will no longer work.

To that end, the FindGL logic focused on the X11 OpenGL libraries has
been relocated to src/libdm/glx and trimmed down for its specific
purpose.  Its results variables have been prefixed with an "X" to avoid
colliding with the standard OpenGL search results.

A more standard FindOpenGL.cmake replaces the FindGL.cmake file for
scenarios where we want to use those libraries.  Need to figure out how
to handle libdmgl in this context - probably need to build dm-gl.c into
the ogl backend rather than linking libdmgl so we don't end up trying to
mix X11 and Apple OpenGL libraries...

While we're at it, move the OpenGL specific bits from dm_util into
libdmgl.