Manage s_fullpath as user data
The only librt data type active here appears to be the db_fullpath.
Rather than making it an intrinsic property of the scene object, have it
be data attached to the object by higher level libraries.

This will set the stage for moving this data type lower in the library
stack.