Break bview functionality out from libdm
libged isn't supposed to depend on libdm, but a couple more uses crept
in.  libged does manipulate views and view information, so handle this
with a separation between view concepts and display concepts.

The screengrab command is more problematic, since the point is to
capture a display rendering.  We'll leverage the new libged plugin
architecture for that case and link just the screengrab plugin to libdm.

There is potential utility in a "dm" command like that defined in MGED
being expressed in a way easily usable by other programs, so one option
here might be to turn the current screengrab plugin into a "dm" plugin
which provides screengrab as one of its commands (or even as a
subcommand of a toplevel dm command.)  The command would be a no-op if
libged doesn't have a dmp set, and would not tie the whole of libged to
libdm, but would allow convenient expression of display manipulation
concepts as a reusable command.