Checkpoint experiment with QTimer
First stab at implementing the technique for update pausing from
https://www.qtcentre.org/threads/58630-Dock-Widget-Custom-Resizing

Sort of works, but doesn't reliably repaint the window.  Disabled in
this commit as it's not working well enough, but it may be worth
exploring combining the technique with some variant of the signal/slot
setup of the dmglt multithreading widget to see if we can get a reliably
responsive renderer for Mac.  Basically, the idea would be that instead
of a render thread locking and doing the work in the OpenGL context, the
timer would be used instead to avoid rapid expensive redraws. Then, at
the end of the "render" wait, paintGL would be called in the main thread
as in dmgls to actually do the work, but without the
render-from-separate-thread logic that seems to be a problem with OSX.