Rework normal/drawBegin management of HUD matrices
Qt's QOpenGLWidget interacted poorly with the faceplate matrices being
left up between the end of the drawing cycle and the next drawBegin
call.  Also, the identified purpose for dm_normal didn't seem to be
quite correct - it's primary job seems to be preparing the way for
faceplate drawing, with drawBegin cleaning up after it.

Change this to be a set of explicit beginning and ending functions for
HUD (Head-Up Display) mode, and exit HUD mode once we're done with it.
For now leaving the dm_hud_end call in drawBegin, since MGED and
libtclcad haven't been vetted for issues and MGED at least will almost
certainly have some.