Remove the render thread based Qt widget
In addition to not working on the Mac, this approach to the widget still
doesn't like the libdm faceplate renderings.  Remove it for now - if
we're force at some point to try again VC has it in the history, and
it's showing too many issues to be worth tangling with unless/until we
know we have to to make a usable application.