Time the hashing to see if it is causing issues
Testing with the Stanford Lucy model, it's clear that we can't hash the
complete vlist set after every command.  It introduces a 2 second+ delay
while the hash checks the data.

I think the middle ground may be to hash the names of scene objects
(TODO, but probably fast), count vertices in the vlists (to test), and
have the subset of commands that can actually change vlist data without
altering either of those metrics (i.e. a vertex edit on a BoT or moving
a point in a point cloud) set a flag on the scene object that is cleared
before the post-update hash is calculated (that way any setting of the
flag will cause a view hash difference.)  Really don't want to require
an app to have to list all GED commands with knowledge of what they
might do - that's a headache in both libtclcad and MGED.  The GED
command author has the domain knowledge, so they should be the ones to
let the view know there is potential for internal vlist changes.