Test out a struct directory flag indicating change
The most universal indicator of potential geometry change is the lowest
- the dbio layer.  We don't have an event system broadcasting changes
from that level, but there may be another way.

We introduce a flag in struct directory that can be zeroed by the
calling application.  That flag is set to 1 when the lowest level of the
API (db_put_external) performs a write.

It is up to the caller to do the rest - initialize it, read it, reset it
to zero for the next round, etc.  If the app ignores it it is harmless,
but for change detection it is key.  That flag being set gives high level
logic a way to detect potential changes completely independent of any
specific GED command or library logic.

A test indicates we can indeed detect an object change (in this case,
color updating.)  Which of course leads immediately to the next problem
- we now need to figure out how scene objects get updated after edit in
MGED and Archer, since the color isn't updated currently in qdm in
response to the change...