Start trying to create a data_polygon circle
Yow.  The logic in libtclcad for adding and manipulating various kinds
of polygons is spread over a lot of functions, some of which do part of
the job and then serialize their data into Tcl lists to hand off to
another function, which unpacks them from Tcl and do some more work.

Rather than redo all of the libtclcad code (time consuming and error
prone for no functional improvement), what we'll try to do instead is
extract the essence of the logic into libbview functions for use in new
logic going forward.