Start prototyping expansion of view obj concept
If it can be managed we would prefer to subsume polygons, lines, arrows
and labels under the bview_scene_obj umbrella as view-only objects (and
eventually transient objects such as NIRT shotlines and rtcheck/gqa
overlap visuals.)

libtclcad is managing polygons and the rest very differently, so need to
think about what sort of data containers we're looking for here and how
the routines being translated from tclcad will need to adjust.  May
prove to be too much work to be doable right now, but it feels wrong to
have the polygons be a one-off.  If it is practical it's much easier to
adjust this now rather than having to re-translate it again later, so
let's give it a go.