First cut at 2D view data as scene obj data
If we view lines, arrows, polygons, etc. as particular types of view
objects, the bview_scene_obj container has most of the necessary knobs
already - we just need to enhance it to optionally store the specific
types of information needed for these view elements.

Untested and non-operational thus far - this is just working on what the
containers should be to handle the necessary info.  In principle it
would be better to set up both the data and any processing functions as
pointers/callbacks to allow for fully custom view objects that are user
defined, but I'm not sure yet what that would look like.  Right now
libdm has specific drawing routines for arrows, polygons, etc - if those
populate the s_vlist container of the bview_scene_obj rather than
directly calling the dm routines we might be able to set up a fully
generic mechanism.