Adjust design towards callback
This should in theory allow for an update pass that lets callers define
per-object interaction routines.  Will need to pass mouse current x,y to
the callback, and probably other changes...  thought right now is to
have the callback return code tell libbview that the object vlist has
changed, rather than trying to detect it explicitly.  Then a flag will
be set in the scene obj structure, and that change will update the hash
and trigger view updating.