Add support for arrows to scene obj drawing
This is an experiment, which enables arrow drawing generically across
all vlist based view objects on a per-object basis.  Does the simple
thing and iterates over the bn_vlist, adding an arrow head at the start
of each MOVE command other than the first and at the final DRAW point.

Not sure if this is sensible yet, but it has the advantage that we no
longer require separate line and arrow types.  In theory it also makes
it easy to do things like enabling arrows to show the direction of
polygons, although that is untested.