Successful display of OSMesa rendering in Qt
Add the ability to the qdm test app to use either the Qt OpenGL widget
or the OSMesa software rendering.  The latter leverages libosmesa and
provides a stand-alone OpenGL 2.1 rendering capability that doesn't rely
on a working system OpenGL stack.

Hadn't planned on adding this yet, but my Linux box exhibited some sort
of OpenGL problem preventing both Qt and Tk from putting up an OpenGL
context.  Since that is precisely the situation where this capability is
most needed it seemed worthwhile to take advantage of the inadvertent
testing environment to get this working.  Based on debugging sessions,
had the system OpenGL been working some aspects of this would have "just
worked" and not been correctly set up for software rendering.  Since the
regular context's routines fail, I can tell in this situation when the
OSMesa routines are being called correctly.

Next up is to see if auto-fallback can be arranged if the app starts up,
tries to initialize the OpenGL context, and fails.  Ideally, we would
fire up the software backend automatically in that situation (assuming
we don't just crash... the way qdm is currently failing to work suggests
we may be able to detect the situation and respond...)