Use RGBA output from osmesa rather than RGB
Grr.  Found a working path to get the painting working in Qt, but
feeding the GL_RGB info to QImage didn't work.  Checkpointing the
working path that has been identified, but need to consider.  Up until
now, all the image returns have been returning RGB data with no alpha
channel.  It may be better for consistency to do that here and manually
prepare the input for Qt's consumption.  On the other hand, the alpha
channel has a lot to recommend it from a usability standpoint, and
almost certainly all environments of relevance today will be capable of
returning it. Throwing away the information by assuming RGB output is
also less than ideal.

Options are to switch all image outputs to an RGBA output style, add a
return type that tells the caller what they're getting, or keep
discarding the alpha channel.