Open/close polygons, add restore for editing view
Store the contour's open/closed state at the contour level, rather than
the polygon type.  Also, save the polygon's creation view state so we
can restore it if the editing environment shifts.  (Right now the qdm
autoview calling is interfering with this, but if an object is drawn
first to short-circuit calling autoview the restore works.)