Upgrade Clipper library to version 6.4.2
New development work is motivating using features present in newer
versions of Clipper.  The API changed a bit, so we need to update our
code accordingly.  This should be more extensively tested - it works for
a simple test case, but things like hole introduction will need to be
checked.

In principle this version will tell us if a contour is open or closed
and support operations on polylines in addition to polygons, which is
the immediate motivation for the upgrade.  In order for the default
state of a bg_polygon to remain closed when an array is calloced using
the older libtclcad code, we switch the flag from a set closed flag
(i.e. closed == 1 indicating a closed polygon) to an open flag (open ==
0 now being the closed polygon setting, which will be the normal
default for zero initialized arrays.