Check bbox calculation before drawing
Behavior is a little odd - at the bottom of the screen the text gets
clamped and drawn, but at the top the bbox calculation fails and the
text disappears.  Need to double check how we're handling the bbox
return - probably returning an error when we shouldn't be, or some such,
and the clamping behavior is also a bit odd...  May need to validate the
visibility of the proposed screen coordinates differently.