Was overthinking bbox return behavior
Just pass back what fontstash gives us.  Interestingly, there is some
sort of clamping going on (either in our code or theirs) to avoid the
text being offscreen completely.  That's probably not what we actually
want for this case, so will need to track down where that's coming from.
May want to make it conditional, since there are potential uses cases
(debugging if nothing else) but for this we want the label to stay
associated with its specified 3D space.