Stub in librt primitive functab entry  for labels
Ew. dm_label_primitive is special casing all of the primitives to create
labels.  This really really belongs with each primitive.  Add the slot
(currently non-functional) to librt so we can work on setting up a
version of this that will be cleaner.  Not planning to migrate MGED or
libtclcad codes, so we won't be able to immediately delete
dm_label_primitive or the MGED version, but intent is to set the stage
for a cleaner approach in new code.