strcmp -> bu_strcmp
The logic now seems more or less able to duplicate the results of the
standard who command.  The main difference is that the standard logic
reports the display list entries, where this version builds the list up
from just the displayed solid objects and doesn't need a separate
display list.  One of the consequences is the new output will reflect
when a toplevel model's components aren't displayed - for example,
ktank's who2 output reflects that there are regions in the hull which
are empty and thus can't be drawn.

Need to see what the performance implications are for large models.