WIP function to encapsulate coloring rules
The rules for coloring are embedded in display_list.c and tree walking
in librt.  We'll probably need to track this in the tree walk to avoid
avs loading and unloading more times than needed, but even so this
function (once complete) will be useful for two reasons:  1) it will
allow a caller that ISN'T tree walking to get color info for a fullpath
without having to set up a tree walk, and 2) it will serve as a concise
and clear definition of what the actual sources and precedence are for
object colors.

Need to do some more research into current logic to ensure this function
reflects the rules correctly, but already seems to be coloring moss.g,
ktank.g and m35.g as expected in quick testing.