Avoid an osmesa crash involving point drawing
I'd like to understand this better, but the following sequence of
commands crashed libosmesa using an experimental view zapping feature:

draw2 component
view2 obj s1 polygon create 330 220 sq
view2 obj s1 update 740 625
Z -v
view2 obj s1 polygon create 330 220 sq

When s1 re-created, a[j].inputstride is 16 and vptr->stride is 0.  This
doesn't happen without component drawn,  It does manifest for each view2
obj manipulation of polygon shapes.  If we create a label it seems to
clear the issue.   If we erase2 component we also don't get the issue.

This may indicate there's an issue with our OpenGL state management, but
examining the line in this function where vptr->stride is used, if start
is zero (which is the case when we're doing this sequence) it doesn't
matter what vptr->stride is - 0*vptr->stride == 0.  Since vptr->stride
has no impact, we shouldn't need the assert fail in that case.

This isn't fully satisfactory - if stride being zero is a problem
elsewhere this may just be deferring the problem to further down the
pipeline - but going ahead and committing to document that it does seem
to improve behavior in the above test case.

While this change seems to result in drawing proceeding essentially as
expected, the point color on the s1 recreation appears to be the brown
of component rather than yellow.  Need to check into that to see if it
is expected, or if it is indicative of some issue that might be
manifesting in the assert behavior we're encountering.