Start roughing out a grouping object
As I'm working with this, I'm coming more to understand why there was a
tiered system of display_list objects and struct solid entities.  The
solids in isolation just can't preserve enough semantic meaning to
preserve the original intent of a draw command through geometry
changes.  The data I'm likely to want at a group looks like scene obj
data (since the idea of children was introduced, at any rate) but the
conceptual distinction is more important than I originally thought.

Try to illustrate in the comment why we need bview_scene_group in
addition to bview_scene_obj in the comments - this will manifest
concretely when the draw2/erase2/who2 commands are updated to reflect
the behaviors outlined here, but try to capture the initial thoughts
where the definition lives so the reason for it is more apparent down
the road.