First properly colored shaded moss.g
Yikes.  Obscure but EXTREMELY important setting that sets up some key
OpenGL state for ambientColor and friends in dm_set_fg.  Was seeing
geometry drawing in the scene but it was either black or using the prior
color from previous drawing operations.

Diffing the various drawing behavior state changes as reported by
apitrace (https://github.com/apitrace/apitrace) pointed to GL_AMBIENT,
GL_SPECULAR and GL_DIFFUSE being zero in the incorrect drawing, but set
in the correct.  That was critical to zeroing in on the issue - a very
useful tool to know about for this sort of debugging.