Add a slight rework of bigE mode as draw2 -m3
The current draw command's -m3 uses brep and isn't working right now,
nor is that feature ready for prime time.  The brep command will be the
right place to expose that feature until its much more stable and
performant.  bigE also has its issues, but it's a classic feature of
MGED.  All of the currently enabled shading modes also depend on working
facetization, so it's not really a stretch in that regard.