Grab some size initialization from display_list.c
This sets up s_size for adaptive plotting, but I'm not convinced its the
right way to do it.

We really need to establish both primitive sizes and the basic view
before doing the full adaptive generation (which is why draw has an
autoview invocation in the beginning) but to get an accurate bounding
box from the ft_bbox calls to do an autoview don't we also need to apply
matrices in a tree walk?  May need to plot some of the steps to properly
understand what is happening (and what needs to happen) when.