More work on adaptive plotting with draw2
This logic attempts to be more explicit and systematic about what we're
doing with adaptive plotting - specifically, we need some sense of view
and size information before we produce the wireframes, which is counter
to how autoview is normally handled (autoview looks at what is in the
scene to generate a reasonable view.  Hence we have a classic
chicken/egg problem.)

The resolution is to use either the primitive bounding boxes (when
available) or bounding of the non-adaptive default wireframes to
generate initial estimates of the size, and use that information to set
up the view the way autoview would using the vlists.  Note that final
autoview pass should still be done the standard way to produce reliably
visible results.  (The wireframes should usually be sized similarly to
the primitive bounding boxes, but nothing actually guarantees that.)