Add lcnt subcmd and opts to list db view objs
Quick and dirty addition to view2 command so we can see what the s_vlist
size is during LoD redrawing.  Probably want to report actual line
segments or some such rather than the bu_list_len - the latter is
probably not what the user is expecting - but this serves the immediate
need.