Add target_link_libraries to static for includes
With the advent of imported targets in CMake building, we can't skip the
target_link_libraries step for static builds anymore.  Include
directories, compile definitions, etc. from those targets will be needed
to compile the libraries.

Hit this with Qt when trying to do a BRLCAD_ADDLIB build using the
imported targets - include dirs and definitions are coming in from those
targets that we need to successfully build.